dev/lol_plandex/requirements.txt
aiohttp
asyncpg
cassiopeia
fastapi
pandas
psycopg2
pytest
pytest-asyncio
python-dotenv
sqlalchemy
uvicorn